Use ExecutorService
The ExecutorService is now used to process tasks. One task is consisting
of prompting the user, reading the user's input and processing the
command. At the end of processing the command a new task is submitted
which again consists of promting the user... and so on. On quit command
the executor is shutdown and the task executing the quit command is the
last processed one.